feat!: disable flake_ref module
