+++ crates/runix/src/lib.rs
@@ -203,7 +203,7 @@ use async_trait::async_trait;
 pub mod arguments;
 pub mod command;
 pub mod command_line;
-pub mod flake_ref;
+// pub mod flake_ref;
 pub mod flake_ref2;
 pub mod installable;
 pub mod registry;
